requirements.txt
appdirs==1.4.3
asgiref==3.2.3
cfgv==3.1.0
distlib==0.3.0
Django==2.2.11
django-auditlog==0.4.7
django-crum==0.7.5
django-dotenv==1.4.2
django-debug-toolbar==2.2
django-fsm==2.7.0
django-fsm-admin==1.2.4
django-jsonfield==1.4.0
filelock==3.0.12
identify==1.4.11
nodeenv==1.3.5
psycopg2-binary==2.8.4
python-dateutil==2.6.0
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.1
django-import-export==2.0.2
toml==0.10.0
boto==2.49.0
django-ses==0.8.14
django-daterange-filter==1.3.0
django-fsm-log==1.6.2
django-su==0.9.0
https://github.com/litchfield/django-liststyle/archive/master.zip
django-phonenumber-field==4.0.0
phonenumbers==8.11.5
https://github.com/acastellana/django-admin-comments/archive/patch-1.zip
django-money==1.0
django-hashid-field==3.1.1
django-extensions==2.2.9
django-health-check==3.12.1
psutil==5.7.0
django-cors-headers==3.2.1
django-settings-export==1.2.1
django-tastypie==0.14.3
